fix(package-name): don't assume the package directory is the package name
